+++ Gastrox/Views/PohybyView.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace Gastrox.Views;
+
+public partial class PohybyView : UserControl
+{
+    public PohybyView()
+    {
+        InitializeComponent();
+    }
+}